Database Lookup Complete
Database lookup function is complete. It serves as a GUI for scans on the transactions table. User field searches for the same ID appearing either in the buyer or seller column, therefore when it has contents, buyer and seller are disabled. When buyer or seller have contents, user is disabled. User will always take priority over buyer and seller in a scan. Also removed some imports not in user.